- Fixed farplane not starting at 500 - Fixed a bug in the Unity Fbx Exporter that was causing models to get way bigger than supposed to. Gains in size of up to 10x. Performance is way faster now. - Added PerMaterial option to asset gathering, gets 40% smaller scene sizes than Brute force.

+++ unity/Project/JanusExporter/Assets/JanusExporter/Editor/Data/Room/Room.cs
@@ -9,7 +9,7 @@ namespace JanusVR
     public class Room
     {
         [XmlAttribute("far_dist")]
-        public float? FarDistance { get; set; }
+        public int? FarDistance { get; set; }
 
         [XmlElement(typeof(RoomObject), ElementName = "Object")]
         [XmlElement(typeof(LinkObject))]
